+++ src/console/controllers/MigrateController.php
@@ -366,6 +366,10 @@ class MigrateController extends BaseMigrateController
                     $this->stdout(PHP_EOL . "$applied from $total " . ($applied === 1 ? 'migration was' : 'migrations were') . ' applied.' . PHP_EOL, Console::FG_RED);
                     $this->stdout(PHP_EOL . 'Migration failed. The rest of the migrations are canceled.' . PHP_EOL, Console::FG_RED);
                     Craft::$app->disableMaintenanceMode();
+                    Craft::$app->getProjectConfig()->reset();
+                    if (!$this->restore()) {
+                        $this->stdout("\nRestore a database backup before trying again.\n", Console::FG_RED);
+                    }
                     return ExitCode::UNSPECIFIED_ERROR;
                 }
                 $applied++;
@@ -416,6 +420,14 @@ class MigrateController extends BaseMigrateController
         }
 
         $res = parent::actionUp($limit);
+
+        if ($res === ExitCode::UNSPECIFIED_ERROR) {
+            Craft::$app->getProjectConfig()->reset();
+            if (!$this->restore()) {
+                $this->stdout("\nRestore a database backup before trying again.\n", Console::FG_RED);
+            }
+            return $res;
+        }
 
         if ($res === ExitCode::OK && empty($this->getNewMigrations())) {
             // Update any schema versions.