+++ src/console/controllers/UpController.php
@@ -56,9 +56,10 @@ class UpController extends Controller
             $writeYamlAutomatically = $projectConfig->writeYamlAutomatically;
 
             // Craft + plugin migrations
-            if ($this->run('migrate/all', ['noContent' => true]) !== ExitCode::OK) {
-                $this->stderr("\nAborting remaining tasks.\n", Console::FG_RED);
-                throw new OperationAbortedException();
+            $res = $this->run('migrate/all', ['noContent' => true]);
+            if ($res !== ExitCode::OK) {
+                $this->stderr("\nAborting remaining tasks.\n", Console::FG_YELLOW);
+                return $res;
             }
             $this->stdout("\n");
 
@@ -68,15 +69,17 @@ class UpController extends Controller
 
             // Project Config
             if ($pendingChanges) {
-                if ($this->run('project-config/apply') !== ExitCode::OK) {
-                    throw new OperationAbortedException();
+                $res = $this->run('project-config/apply');
+                if ($res !== ExitCode::OK) {
+                    return $res;
                 }
                 $this->stdout("\n");
             }
 
-            // Content migrations
-            if ($this->run('migrate/up', ['track' => MigrationManager::TRACK_CONTENT]) !== ExitCode::OK) {
-                throw new OperationAbortedException();
+            // Content migration
+            $res = $this->run('migrate/up', ['track' => MigrationManager::TRACK_CONTENT]);
+            if ($res !== ExitCode::OK) {
+                return $res;
             }
             $this->stdout("\n");
         } catch (Throwable $e) {