+++ src/elements/GlobalSet.php
@@ -9,6 +9,7 @@ namespace craft\elements;
 
 use Craft;
 use craft\base\Element;
+use craft\base\FieldLayoutProviderInterface;
 use craft\behaviors\FieldLayoutBehavior;
 use craft\elements\db\GlobalSetQuery;
 use craft\helpers\UrlHelper;
@@ -25,7 +26,7 @@ use yii\base\InvalidConfigException;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.0.0
  */
-class GlobalSet extends Element
+class GlobalSet extends Element implements FieldLayoutProviderInterface
 {
     /**
      * @since 4.4.6
@@ -234,7 +235,7 @@ class GlobalSet extends Element
     /**
      * @inheritdoc
      */
-    public function getFieldLayout(): ?FieldLayout
+    public function getFieldLayout(): FieldLayout
     {
         /** @var FieldLayoutBehavior $behavior */
         $behavior = $this->getBehavior('fieldLayout');