+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Tags fields now only show up to five suggestions. ([#13322](https://github.com/craftcms/cms/issues/13322))
+- The `up`, `migrate/up`, and `migrate/all` commands now revert any project config changes created by migrations on failure.
+- The `up`, `migrate/up`, and `migrate/all` commands now prompt to restore the backup created at the outset of the command, or recommend restoring a backup, on failure.
+- Added `craft\console\controllers\BackupTrait::restore()`.
 - Fixed a bug where Single entries weren’t getting preloaded for template macros, if the template body wasn‘t rendered. ([#13312](https://github.com/craftcms/cms/issues/13312))
 - Fixed a bug where asset folders could get dynamically created for elements with temporary slugs. ([#13311](https://github.com/craftcms/cms/issues/13311))